adding snowflake id (WIP)

+++ internal/events/dm.go
@@ -1,7 +1,7 @@
 package events
 
 type Dm struct {
-	DmId      int       `json:"id"`
+	DmId      int64     `json:"id"`
 	UserInfo  User      `json:"userInfo"`
 	UnreadMsg UnreadMsg `json:"unreadMsg"`
 }
